Fix app JSPM client baseUrl
The "file://" URL handler requires the full path to the html file. Before
this commit, this was accomplished by modifying the jspm-generated
config.js file. It was necessary to do it every time jspm init/install was
run, as config.js does not support dynamic values.

This commit fixes this issue by setting the client baseUrl to "." in the
jspm init wizard. As a result, client baseUrl is preserved by jspm.